+++ unstructured_ingest/processes/connectors/sql/singlestore.py
@@ -62,8 +62,10 @@ class SingleStoreConnectionConfig(SQLConnectionConfig):
         try:
             yield connection
         finally:
-            connection.commit()
-            connection.close()
+            try:
+                connection.commit()
+            finally:
+                connection.close()
 
     @contextmanager
     def get_cursor(self) -> Generator["SingleStoreCursor", None, None]: