+++ unstructured_ingest/processes/connectors/sql/snowflake.py
@@ -96,8 +96,10 @@ class SnowflakeConnectionConfig(SQLConnectionConfig):
         try:
             yield connection
         finally:
-            connection.commit()
-            connection.close()
+            try:
+                connection.commit()
+            finally:
+                connection.close()
 
     @contextmanager
     def get_cursor(self) -> Generator["SnowflakeCursor", None, None]: