+++ unstructured_ingest/processes/connectors/sql/sqlite.py
@@ -66,8 +66,10 @@ class SQLiteConnectionConfig(SQLConnectionConfig):
         try:
             yield connection
         finally:
-            connection.commit()
-            connection.close()
+            try:
+                connection.commit()
+            finally:
+                connection.close()
 
     @contextmanager
     def get_cursor(self) -> Generator["SqliteCursor", None, None]: